+++ R/dist_multinomial.R
@@ -5,12 +5,12 @@
 #' The multinomial distribution is a generalization of the binomial
 #' distribution to multiple categories. It is perhaps easiest to think
 #' that we first extend a [dist_bernoulli()] distribution to include more
-#' than two categories, resulting in a categorical distribution.
+#' than two categories, resulting in a [dist_categorical()] distribution.
 #' We then extend repeat the Categorical experiment several (\eqn{n})
 #' times.
 #'
-#' @inheritParams stats::dmultinom
-#'
+#' @param size The number of draws from the Categorical distribution.
+#' @param prob The probability of an event occurring from each draw.
 #'
 #' @details
 #'